# revert scheduler change