Current:
```
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py
***** model tensor in     25.82 ms
***** model schedule in  238.27 ms
***** model opts(32) in  126.48 ms
***** model lower in      56.57 ms
***** model rewrite in   442.46 ms
***** model linearize in 166.69 ms
20200
all 1057.31 ms
```

With revert
```
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py 
***** model tensor in     25.90 ms
***** model schedule in   77.48 ms
***** model opts(32) in  130.81 ms
***** model lower in      57.11 ms
***** model rewrite in   439.47 ms
***** model linearize in 166.56 ms
20200
all 898.34 ms
```

10% slowdown is fine, but 4x isn't