@geohot 2025-02-11
It's still creating all the buffers and running the old break sched. I'm okay with up to 20% slower on that `external_benchmark_schedule` to merge again.