@geohot 2025-02-11
![Screenshot 2025-02-11 at 11 30 11 AM](https://github.com/user-attachments/assets/5b2a5861-cadf-4f32-b29d-82d63dec87b2)

Does look cool though