@Qazalin 2025-02-11
yea the viz of this style of sched_sink is nice, I'll find the diff that gets us the <20%.